services/auth-service/pyproject.toml
[tool.poetry]
name = "auth-service"
version = "1.0.0"
description = "Authentication microservice for freelancer marketplace"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.104.0"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
sqlalchemy = "^2.0.23"
aiomysql = "^0.2.0"
pydantic = {extras = ["email"], version = "^2.12.2"}
pydantic-settings = "^2.1.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
bcrypt = "^4.1.1"
aio-pika = "^9.3.0"
typing-extensions = "^4.0.0"
async-exit-stack = "^1.0.1"
alembic = "^1.12.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
pytest-mock = "^3.12.0"
pytest-cov = "^4.1.0"
httpx = "^0.25.1"
mypy = "^1.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"